Actualización libft, con calloc y strdup

+++ Fundacion42/Cursus/libft/libft/ft_strncmp.c
@@ -6,14 +6,13 @@
 /*   By: dmena-ro <dmena-ro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/04 18:31:35 by dmena-ro          #+#    #+#             */
-/*   Updated: 2022/12/09 21:40:19 by dmena-ro         ###   ########.fr       */
+/*   Updated: 2022/12/14 19:39:21 by dmena-ro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
 #include "libft.h"
 
-size_t	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	size_t	i;
 
@@ -22,7 +21,7 @@ size_t	ft_strncmp(const char *s1, const char *s2, size_t n)
 		return (0);
 	while (s1[i] != '\0' && s2[i] != '\0' && s1[i] == s2[i])
 	{
-		if (i < (n -1))
+		if (i < (n - 1))
 			i++;
 		else
 			return (0);
